requirements.txt
langchain==0.0.170
python-dotenv==1.0.0
streamlit==1.21.0
openai